fix(coderd): address chat agent review follow-ups
