+++ coderd/x/chatd/internal/agentselect/agentselect.go
@@ -8,22 +8,30 @@ import (
 	"golang.org/x/xerrors"
 
 	"github.com/coder/coder/v2/coderd/database"
-	"github.com/coder/coder/v2/coderd/x/chatd/chatagent"
 )
 
-// SelectChatAgent picks the best workspace agent for a chat session from
-// the provided candidates. It applies these rules in order:
+// Suffix marks chat-designated agents during the current PoC. This naming
+// convention is an implementation detail, not a stable contract.
+const Suffix = "-coderd-chat"
+
+// IsChatAgent reports whether name uses the chat-agent suffix convention.
+func IsChatAgent(name string) bool {
+	return strings.HasSuffix(strings.ToLower(name), Suffix)
+}
+
+// FindChatAgent picks the best workspace agent for a chat session from the
+// provided candidates. It applies these rules in order:
 //  1. Filter to root agents only (ParentID is null).
-//  2. Sort stably and deterministically by DisplayOrder ASC, then
-//     Name ASC (case-insensitive), then Name ASC, then ID ASC.
-//  3. If exactly one root agent name ends with chatagent.Suffix
-//     (case-insensitive), return it.
-//  4. If zero root agents match the suffix, return the first root agent
-//     after sorting (deterministic fallback).
-//  5. If more than one root agent matches the suffix, return an error
-//     with an actionable message.
+//  2. Sort stably and deterministically by DisplayOrder ASC, then Name ASC
+//     (case-insensitive), then Name ASC, then ID ASC.
+//  3. If exactly one root agent name ends with Suffix (case-insensitive),
+//     return it.
+//  4. If zero root agents match the suffix, return the first root agent after
+//     sorting (deterministic fallback).
+//  5. If more than one root agent matches the suffix, return an error with an
+//     actionable message.
 //  6. If no root agents exist at all, return an error.
-func SelectChatAgent(
+func FindChatAgent(
 	agents []database.WorkspaceAgent,
 ) (database.WorkspaceAgent, error) {
 	rootAgents := make([]database.WorkspaceAgent, 0, len(agents))
@@ -33,7 +41,7 @@ func SelectChatAgent(
 			continue
 		}
 		rootAgents = append(rootAgents, agent)
-		if chatagent.IsChatAgent(agent.Name) {
+		if IsChatAgent(agent.Name) {
 			matchingAgents = append(matchingAgents, agent)
 		}
 	}
@@ -71,7 +79,7 @@ func SelectChatAgent(
 		}
 		return database.WorkspaceAgent{}, xerrors.Errorf(
 			"multiple agents match the chat suffix %q: %s; only one agent should use this suffix",
-			chatagent.Suffix,
+			Suffix,
 			strings.Join(names, ", "),
 		)
 	}
